Moved functions from musicUtil to discordUtil
It doesnt make much sense for these functions to be in a file called musicUtil. Moved them to a more aptly named file called discordUtil.

Functions moved:
initEmoji
getTarget
getTargetMember
