+++ src/commands/admin/block.ts
@@ -2,8 +2,8 @@ import { Message } from 'discord.js';
 import { logger } from '../../app';
 import { ICommand } from '../../classes/Command';
 import { AddBlocked, blockedUsers } from '../../db/dbBlocked';
-import { getTarget } from '../../util/musicUtil';
 import { createFooter, QuickEmbed, wrap } from '../../util/styleUtil';
+import { getTarget } from '../../util/discordUtil';
 
 export const command: ICommand = {
    name: 'Block',
