+++ src/commands/server/avatar.ts
@@ -1,7 +1,7 @@
 import { MessageEmbed } from 'discord.js';
 import { ICommand } from '../../classes/Command';
-import { getTarget } from '../../util/musicUtil';
 import { createFooter, QuickEmbed } from '../../util/styleUtil';
+import { getTarget } from '../../util/discordUtil';
 
 export const command: ICommand = {
    name: 'avatar',
